0331-1/requirements.txt
easyocr>=1.7
opencv-python>=4.8
pydantic>=2.0
tqdm>=4.0
pyyaml>=6.0
numpy>=1.24
